requirements.txt
colorama==0.4.4
loguru==0.7.0
python-decouple==3.8
tabulate==0.9.0

requests~=2.31.0
pyfiglet~=0.8.post1